README.md
This project is designed for unlimited image conversion to WebP format. With this project, you can easily convert all image files in a folder to WebP format with just one click. This project does not require any special software installation on your computer and can be easily executed in Google Colab.